Initial commit. Still ugly.
This is a port of the Python version of this tool. That one used
matplotlib to generate graphs. I'm hoping to use Javascript/JQuery to
generate something more interactive. There's still a lot to do.

As it currently stands, 50% of the functionality has been ported. The
script ingests Markdown text, converts to HTML, strips to plaintext, can
calculates wordlength and reader velocity. The graphing features haven't
been ported yet. The plan is to use jqplot.

There's also quite a bit of clean-up to do. Right now, all of the
development is happening in a single HTML file. Ultimately, this should
be split into one or more .js files. Also, I'm currently using CDN
variants of jQuery and Markdown-It. I should probably have those local.